add serde package and change storage to contain compatability table.
